The FlowTable class will now respond to UDP and ICMP.
UDP and ICMP flows identified by the DPAE will not cause the
controller application to display an error. It should now be
possible to put flow treatment on UDP flows. ICMP flows will not
be treated. The treatment of flows is out of scope for my work,
this extension should provide help in the future however.